feat : alert 라이브러리

+++ frontend/client/src/components/login/loginForm.jsx
@@ -1,22 +1,46 @@
 import React, { useState } from 'react';
 import axios from 'axios';
-import { useNavigate } from "react-router-dom";
+import { useNavigate } from 'react-router-dom';
 import { Button, Label, TextInput } from 'flowbite-react';
+import Swal from 'sweetalert2';
+
 const LoginForm = () => {
   const [id, setId] = useState("");
   const [pass, setPass] = useState("");
   const navigate = useNavigate();
   const requsetLogin = () => {
+    if(id.length  === 0){
+      Swal.fire({
+      "text" : '아이디를 입력하세요',
+      "confirmButtonColor" : '#3085d6'
+    });
+      return;
+    }
+
+    if(pass.length === 0){
+      Swal.fire({
+        "text" : '비밀번호를 입력하세요',
+        "confirmButtonColor" : '#3085d6'
+      });
+      return;
+    }
+
     axios.post('https://spring.secretzoo.site/api/auth/login',
       {
         "userId": id,
         "password": pass,
       }
-    ).then((response) => {
+    ).then(response => {
       sessionStorage.setItem('authorization', response.headers['authorization']);
       sessionStorage.setItem('refresh_token', response.headers['refresh_token']);
       sessionStorage.setItem('user', response.data);
       navigate('lobby');
+    }).catch(e => {
+      Swal.fire({
+        "text" : '아이디 혹은 비밀번호가 일치하지 않습니다',
+        "confirmButtonColor" : '#3085d6'
+      });
+      return;
     })
   }
 
@@ -44,7 +68,9 @@ const LoginForm = () => {
           <TextInput
             value={pass}
             onChange={(e) => setPass(e.target.value)}
-            type="password" required />
+            type="password" 
+            required
+            placeholder='비밀번호' />
         </div>
         <Button type="submit" onClick={(e) => { e.preventDefault(); requsetLogin() }}>로그인</Button>
       </form>
